trying without separate server
