+++ loadtest/tests/index.js
@@ -3,7 +3,7 @@ require('dotenv').config();
 const action = require('../lib/action');
 
 const {
-  metricsServer,
+  // metricsServer,
   audio,
   chat,
   note,
@@ -18,7 +18,7 @@ const logger = require('../lib/logger');
 const run = require('../lib/run');
 
 module.exports = {
-  metricsServer,
+  // metricsServer,
   audio,
   chat,
   note,
